requirements.txt
awesome-slugify
configargparse
igv_plotter
peewee
pysam
pyvcf
